requirements.txt
streamlit
langchain
sentence-transformers
psycopg2-binary
pgvector
ollama
beautifulsoup4
langchain-text-splitters